+++ CopilotKit/packages/react-ui/src/types/css.ts
@@ -4,12 +4,13 @@ import { CSSProperties } from "react";
 export interface CopilotKitCSSProperties extends CSSProperties {
   "--copilot-kit-primary-color"?: string;
   "--copilot-kit-contrast-color"?: string;
+  "--copilot-kit-background-color"?: string;
+  "--copilot-kit-input-background-color"?: string;
   "--copilot-kit-secondary-color"?: string;
   "--copilot-kit-secondary-contrast-color"?: string;
-  "--copilot-kit-background-color"?: string;
-  "--copilot-kit-muted-color"?: string;
   "--copilot-kit-separator-color"?: string;
-  "--copilot-kit-scrollbar-color"?: string;
-  "--copilot-kit-response-button-color"?: string;
-  "--copilot-kit-response-button-background-color"?: string;
+  "--copilot-kit-muted-color"?: string;
+  "--copilot-kit-shadow-sm"?: string;
+  "--copilot-kit-shadow-md"?: string;
+  "--copilot-kit-shadow-lg"?: string;
 }