+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,34 @@
 # ui
 
+## 1.8.0-next.4
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.0-next.4
+- @copilotkit/shared@1.8.0-next.4
+
+## 1.8.0-next.3
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.0-next.3
+- @copilotkit/shared@1.8.0-next.3
+
+## 1.7.2-next.2
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.7.2-next.2
+- @copilotkit/shared@1.7.2-next.2
+
+## 1.7.2-next.1
+
+### Patch Changes
+
+- 73f5eaa: - fix(react-core): export missing action-related types for public API
+  - @copilotkit/runtime-client-gql@1.7.2-next.1
+  - @copilotkit/shared@1.7.2-next.1
+
 ## 1.7.2-next.0
 
 ### Patch Changes