+++ CopilotKit/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,38 @@
 # ui
 
+## 1.8.0-next.4
+
+### Patch Changes
+
+- @copilotkit/react-core@1.8.0-next.4
+- @copilotkit/runtime-client-gql@1.8.0-next.4
+- @copilotkit/shared@1.8.0-next.4
+
+## 1.8.0-next.3
+
+### Patch Changes
+
+- @copilotkit/react-core@1.8.0-next.3
+- @copilotkit/runtime-client-gql@1.8.0-next.3
+- @copilotkit/shared@1.8.0-next.3
+
+## 1.7.2-next.2
+
+### Patch Changes
+
+- @copilotkit/react-core@1.7.2-next.2
+- @copilotkit/runtime-client-gql@1.7.2-next.2
+- @copilotkit/shared@1.7.2-next.2
+
+## 1.7.2-next.1
+
+### Patch Changes
+
+- Updated dependencies [73f5eaa]
+  - @copilotkit/react-core@1.7.2-next.1
+  - @copilotkit/runtime-client-gql@1.7.2-next.1
+  - @copilotkit/shared@1.7.2-next.1
+
 ## 1.7.2-next.0
 
 ### Patch Changes